[Postgres] removed tablespace from Docker container setup.
Tablespace setup has been removed from the Docker container due to file permission complications. This issue arises because the database user account, postgres, requires permissions to read and write to the tablespace directory.

When running a dedicated database instance this can be achieved through chowning the tablespace directory for the postgres user account.

To mitigate this issue in Docker, a directory on the host machine can be mounted with read and write permissions for the all user group. This poses security issues.