[Postgres] Switched pgtype.Numeric to decimal.Decimal.
There are inherent issues with floating point precision.

pgtype.Numeric is a basic container for Postgres data and does not support arithmetic and comparison operations. This necessitates the conversion to float64 for any such operations which can result in precision loss.

The decimal.Decimal library offers accurate representation of decimal numbers with precise operations and comparisons. It also offers the much needed round half to even, also known as Banker's Rounding, which is essential when working with money.