[Postgres] Interface for External Integrations.
An interface for external integrations with the Postgres database has been created. This will assist with the mocking of all queries, including the transactions, in packages leveraging the database. Base custom errors have also been added for use with the external Postgres interface to abstract away the internal errors.

The Querier interface and methods generated by SQLC are now private to the Postgres package.