[Auth] Authorization and Encryption Methods.
Methods to generate, verify, and re-issue JSON Web Tokens. The JWT will be encrypted and have a validity period that can be defined in the services configuration file. The encrypted token will additionally contain the Client ID.

Other methods to encrypt to and from Base64 encoded strings are also included. These methods will be utilized to encrypt the keys for exchange rate offers. These offers will be stored in the Redis cache with a TTL of a few minutes.