[REST] Gin Testing Concurrency Safe.
The Gin instance in the test suite led to a race condition with setting the mode to test.

Additionally, each test grid's test cases would add paths to the tests Gin instance in parallel. This would lead to another race condition on the path struct within Gin. The solution is to either refactor the tests to allow each test to run its own instance of Gin for testing or not to run tests in parallel.

After profiling, I was unable to register a significant performance hit in regard to memory from having each test create and use its own Gin instance. Each instance adds a memory overhead of under 512kb and memory is cheap and plentiful. Serializing the tests could lead to longer wait times during development.