[REST] Fiat Exchange Endpoints. (#40)
Endpoints have been established to obtain an exchange rate offer and then to execute said transfer offer.

* Offer /fiat/exchange/offer:
Submit a request for a price quote.
The source and destination currencies must be supplied. The amount to be deposited into the destination account will be automatically calculated based on the exchange rate and the source amount supplied in the request. The request will only be valid for a limited period of time, and the Unix expiration deadline will be returned. The Offer ID is the encrypted Offer ID used to index the offer in the Redis cache.

* Transfer /fiat/exchange/transferr:
Complete a request for a Fiat transfer.
The encrypted Offer ID from the Offer endpoint will be supplied as a parameter to this request. If the Offer is valid and there are sufficient funds, it will be completed.