[Auth] Checks for Deleted User Accounts.
There was a bug where a user could delete their account but use a valid JWT to continue API requests. Checks have been added to the authentication middleware for both REST and GraphQL for deleted accounts.

The options considered :
1. A table join of the users table with the requested data table.
2. Execute a single column query against the users table to retrieve the deleted status in the authentication middleware.

Option 2 was selected due to a similar performance impact and to maintain a separation of concerns between the microservices packages. This PR will bring the release version to v1.1.1.